requirements.txt
Django==2.2.4
django-cors-headers==2.4.0
djangorestframework==3.9.3
django-filter==2.0
django-enumfields==0.10.0
django-healthz==0.0.5
django-rehive-extras==0.0.1
drf-rehive-extras==0.0.3
drf-yasg==1.15.0
gunicorn==19.9.0
requests==2.20.0
psycopg2==2.7.5
rehive==1.2.5
sentry-sdk==0.6.9
stripe==2.41.0
